ci(unittest): partial coverage of many to any
