+++ NineteenSevenFour.Testing.Example.Domain/NineteenSevenFour.Testing.Example.Domain.csproj
@@ -19,4 +19,8 @@
     <PackageReference Include="AutoBogus" Version="2.13.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\NineteenSevenFour.Testing.FluentBogus.Relation\NineteenSevenFour.Testing.FluentBogus.Relation.csproj" />
+  </ItemGroup>
+
 </Project>
